Remove quotation marks around fu-types from PARAMS.sunny_cove
With the quotation marks, the functionality of the first function unit cannot be parsed correctly because the quotation mark is considered to be part of the numerical string. The result of converting such "numerical" string is 0, which is then considered as full capability (that can process all kinds of ops).